issues/4 refactor(index): split functions
